% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_year_group.R
\name{add_year_group}
\alias{add_year_group}
\title{Add year groups to a data frame}
\usage{
add_year_group(df, year_col = year)
}
\arguments{
\item{df}{A data frame.}

\item{year_col}{Unquoted column that contains the year (numeric/integer).}
}
\value{
\code{df} with an added character column \code{year_group}.
}
\description{
Adds a \code{year_group} column based on fixed ranges.
}
\examples{
df <- data.frame(year = 2012:2023)
add_year_group(df, year)
}
