% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{array_draw_data}
\alias{array_draw_data}
\title{Simulate data by sampling without replacement from a space-time grid}
\usage{
array_draw_data(sim_array, n_data, N)
}
\arguments{
\item{x}{TODO}
}
\description{
Given a space-time grid produced by
  \code{sim_spacetime_spectral()}, draw sampling locations by replacement.
  Then draw positive sample numbers given a defined sample size.
}
