% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{build_africa_bbox_and_crop}
\alias{build_africa_bbox_and_crop}
\title{Build a buffered Africa bbox, cropped shapefiles, and mask}
\usage{
build_africa_bbox_and_crop(
  africa_sf,
  water_sf,
  buf_km = 200,
  crs_ll = 4326,
  crs_metric = 3857
)
}
\arguments{
\item{africa_sf}{An `sf` object containing Africa land polygons
(e.g. admin0 or admin1 for the African continent).}

\item{water_sf}{An `sf` object containing water bodies for Africa.}

\item{buf_km}{Numeric. Buffer (in kilometers) by which to expand the
Africa bounding box in all directions. Default is 200.}

\item{crs_ll}{Integer or `crs` object. The longitude/latitude CRS
used for plotting (default: EPSG 4326).}

\item{crs_metric}{Integer or `crs` object. A planar CRS used for
buffering and intersections (default: EPSG 3857).}
}
\value{
A list with components:
\describe{
  \item{xlim}{Length-2 numeric vector of longitude limits (xmin, xmax).}
  \item{ylim}{Length-2 numeric vector of latitude limits (ymin, ymax).}
  \item{bbox_sfc_ll}{An `sfc` polygon giving the buffered bbox in lon/lat.}
  \item{africa_crop}{Cropped `sf` object of Africa polygons.}
  \item{water_crop}{Cropped `sf` object of water bodies.}
  \item{africa_mask}{Unioned `sf` polygon mask for Africa (for raster masking).}
}
}
\description{
This helper constructs a bounding box around the African continent
(optionally buffered by a given distance in km), crops the supplied
Africa and water `sf` layers to that box, and returns a union mask
plus the x/y limits for plotting.
}
\examples{
\dontrun{
  bbox_info <- build_africa_bbox_and_crop(shape_Africa, shape_water)
  xlim <- bbox_info$xlim
  ylim <- bbox_info$ylim
  africa_mask <- bbox_info$africa_mask
}

}
