% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{emplogit}
\alias{emplogit}
\title{Empirical logit function}
\usage{
emplogit(x, N)
}
\arguments{
\item{x}{TODO}
}
\description{
The empirical logit is a version of the logit transform that is
  shrunk towards a value, in our case 0.5. This avoids issues with infinite
  numbers when prevalence is 0% or 100%.
}
