% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variogram.R
\name{get_vario_cor}
\alias{get_vario_cor}
\title{TODO}
\usage{
get_vario_cor(
  dist_space,
  dist_time,
  length_space,
  length_time,
  kernel_space_power = 1,
  kernel_time_power = 2
)
}
\arguments{
\item{x}{TODO}
}
\description{
Get the correlation from a spatial-temporal variogram model.
TODO - more text.
}
