% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_vario_loglike}
\alias{get_vario_loglike}
\title{TODO}
\usage{
get_vario_loglike(
  dist_space,
  dist_time,
  dist_z,
  length_space,
  length_time,
  nugget,
  partial_sill,
  kernel_space_power = 1,
  kernel_time_power = 2
)
}
\arguments{
\item{x}{TODO}
}
\description{
Get the likelihood from a spatial-temporal variogram model.
TODO - more text.
}
