% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{lonlat_to_cartesian}
\alias{lonlat_to_cartesian}
\title{Convert Latitude/Longitude to Local Cartesian Coordinates}
\usage{
lonlat_to_cartesian(
  lon,
  lat,
  lon_centre = mean(range(lon)),
  lat_centre = mean(range(lat))
)
}
\arguments{
\item{lon}{Numeric vector of longitudes in decimal degrees.}

\item{lat}{Numeric vector of latitudes in decimal degrees.}

\item{lon_centre}{Longitude (in degrees) for the centre of the projection.}

\item{lat_centre}{Latitude (in degrees) for the centre of the projection.}
}
\value{
A list with elements `x` and `y` representing Cartesian coordinates
  in kilometres. These elements will mirror the structure of the inputs, for
  example if `lon` and `lat` are matrices then these will be matrices
}
\description{
Converts geographic coordinates (latitude, longitude) to
Cartesian x and y coordinates in kilometres using a local tangent plane 
(equirectangular projection) centred on a specified point. 
This assumes a flat Earth approximation.
}
