% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prev_colors.R
\name{prev_colors}
\alias{prev_colors}
\title{Continuous prevalence color palette}
\usage{
prev_colors()
}
\value{
A character vector of hex color codes and named colors defining
  the prevalence color scale, ordered from low to high prevalence.
}
\description{
Returns the color palette used for continuous prevalence maps.
}
\examples{
pp_cols()

}
