% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prev_vals.R
\name{prev_vals}
\alias{prev_vals}
\title{Rescaled prevalence break values for plotting}
\usage{
prev_vals(breaks = c(0, 1, 5, 10, 20, 30, 50, 70, 90, 95, 100))
}
\arguments{
\item{breaks}{Numeric vector of prevalence percentages.
Defaults to \code{c(0, 1, 5, 10, 20, 30, 50, 70, 90, 95, 100)}.}
}
\value{
A numeric vector of rescaled values in \eqn{[0,1]}.
}
\description{
Returns rescaled values in \eqn{[0,1]} corresponding to prevalence
percentage breakpoints, suitable for use with
\code{scale_fill_gradientn(values = ...)}.
}
\examples{
pp_vals()

}
