% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{project_to_local_laea}
\alias{project_to_local_laea}
\title{Project points to a local LAEA system centered on the data}
\usage{
project_to_local_laea(
  df,
  lon_col = "longitude",
  lat_col = "latitude",
  crs_ll = 4326
)
}
\arguments{
\item{df}{A data frame with longitude/latitude columns.}

\item{lon_col}{Name of longitude column (string), default "longitude".}

\item{lat_col}{Name of latitude column (string), default "latitude".}

\item{crs_ll}{CRS for lon/lat, default EPSG 4326.}
}
\value{
A list with:
\describe{
  \item{xy_km}{Matrix of projected coordinates in km (columns Xkm, Ykm).}
  \item{crs_laea}{The LAEA proj4 string used.}
  \item{lon0}{Center longitude.}
  \item{lat0}{Center latitude.}
}
}
\description{
Project points to a local LAEA system centered on the data
}
