% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_csv}
\alias{save_csv}
\title{Save CSV (mirrors save_figs style)}
\usage{
save_csv(
  name,
  df,
  data_dir = file.path(here::here(), "R_ignore/R_scripts/outputs/plots"),
  na = "",
  append_date = FALSE,
  gzip = FALSE,
  ...
)
}
\arguments{
\item{name}{Base name of the file (without extension)}

\item{df}{Data frame to write}

\item{data_dir}{Output directory (default: "analysis/data_derived/prev_summary_tables")}

\item{na}{String to use for missing values (default: "")}

\item{append_date}{If TRUE, append YYYY-MM-DD to the file name (default: FALSE)}

\item{gzip}{If TRUE, write a gzipped CSV (.csv.gz) (default: FALSE)}

\item{...}{Other args passed to readr::write_csv()}
}
\value{
(invisible) path to the written file
}
\description{
Save CSV (mirrors save_figs style)
}
