% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_spacetime_spectral}
\alias{sim_spacetime_spectral}
\title{Simulate values on a space-time grid by drawing from a GRF using spectral methods}
\usage{
sim_spacetime_spectral(
  x_range,
  y_range,
  t_range,
  nx,
  ny,
  nt,
  x_buffer = 10,
  y_buffer = 10,
  t_buffer = 10,
  length_space = 1,
  length_time = 1,
  kernel_space_type = "RBF",
  kernel_time_type = "RBF",
  mu = 0,
  sigma = 1
)
}
\arguments{
\item{x}{TODO.}
}
\description{
simulate values on a space-time grid. The range in each
  dimension is set by the user along with the number of grid cells.
  Simulation is via a spectral approximation, which is efficient for large
  grid sizes.
}
